+++ ArpSrc/MDust/MIndexedList.cpp
@@ -9,10 +9,10 @@
  * <URL:http://www.angryredplanet.com/>.
  */
 
-#include <assert.h>
+#include <cassert>
 #include <malloc.h>
-#include <stdio.h>
-#include <math.h>
+#include <cstdio>
+#include <cmath>
 #include <support/ClassInfo.h>
 
 #include "MDust/MIndexedList.h"
@@ -69,7 +69,7 @@ ArpIndexedNode* MIndexedList::HeadNode() const
 }
 
 status_t MIndexedList::ChainHeadFrom(int32 index, ArpIndexedNode **node) {
-	long hash = HashForIndex(index);
+	int32 hash = HashForIndex(index);
 
 	// We used to hash to 1 minus the hash of the index, but we've removed that.
 	// If clients want that, they can send it themselves.
@@ -91,7 +91,7 @@ Print();
 
 int32 MIndexedList::IndexHeadFrom(int32 index)
 {
-	long hash = HashForIndex(index);
+	int32 hash = HashForIndex(index);
 	if (! HashExists(hash)) return B_ERROR;
 	return hash * ChainIndexes();
 }
@@ -224,16 +224,16 @@ void MIndexedList::GrowToAtLeast(int32 countArg)
 	
 	if (chain == 0) {
 		chain = (ArpIndexedNode**)malloc(sizeof(ArpIndexedNode)*count+2);
-		for (long i=0; i<count; i++) chain[i] = 0;
+		for (int32 i=0; i<count; i++) chain[i] = 0;
 
 	} else {
 		ArpIndexedNode	**tmp;
 		tmp = (ArpIndexedNode**)malloc(sizeof(ArpIndexedNode)*count+2);
-		for (long i=0; i<count; i++) tmp[i] = 0;
+		for (int32 i=0; i<count; i++) tmp[i] = 0;
 		if (oldcount < count) {
-			for (long i=0; i<oldcount; i++) tmp[i] = chain[i];
+			for (int32 i=0; i<oldcount; i++) tmp[i] = chain[i];
 		} else {
-			for (long i=0; i<count; i++) tmp[i] = chain[i];
+			for (int32 i=0; i<count; i++) tmp[i] = chain[i];
 		}
 		free(chain);
 		chain = tmp;
@@ -297,7 +297,7 @@ void MIndexedList::CopyTo(MIndexedList *copy)
 {
 	copy->GrowToAtLeast(count);	
 
-	for (long k=0; k<count; k++) {
+	for (int32 k=0; k<count; k++) {
 		if (chain[k] != 0) copy->chain[k] = chain[k]->Copy();
 	}
 }
@@ -322,7 +322,7 @@ void MIndexedList::Print()
 		return;
 	}
 
-	for(long i=0; i<count; i++) {
+	for(int32 i=0; i<count; i++) {
 		printf("Chain %d", (int)i);
 		if (chain[i] == NULL) {
 			printf("\tNULL\n");
@@ -362,11 +362,11 @@ void MIndexedList::SyncTailNode()
 // but in the case of the adding functions, they need to grow to that size.
 int32 MIndexedList::HashForIndex(int32 index)
 {
-	long hash = (long)floor(index / chainIndexes);
+	int32 hash = (int32)floor(index / chainIndexes);
 	if (hash < 0) return(ARP_HASH_ERROR);
 	return(hash);
 
-//	long index = floor( ((ppqnTime / PPQN) / chainIndexes) );
+//	int32 index = floor( ((ppqnTime / PPQN) / chainIndexes) );
 //	if (index < 0) return(ARP_HASH_ERROR);
 //	return(index);
 }
@@ -428,7 +428,7 @@ void MidiAbstractList::VerifyListPosition(MidiAbstractList *lArg) {
 // Check to see if node should be the head of a chain.
 /*
 void MidiAbstractList::KeepChainsConsistent(MidiListNode *node) {
-	long index = IndexForPPQN(node->Time());
+	int32 index = IndexForPPQN(node->Time());
 	if (index >= count)
 		GrowToAtLeast(index + ARP_CHAIN_INCREMENT);	
 
@@ -444,7 +444,7 @@ void MidiAbstractList::KeepChainsConsistent(MidiListNode *node) {
 status_t MIndexedList::NodeBefore(int32 index, ArpIndexedNode **answer,
 						int32 findNode)
 {
-	long hash = HashForIndex(index);
+	int32 hash = HashForIndex(index);
 	if (hash >= count) hash = count - 1;
 	if (! HashExists(hash)) return B_ERROR;
 
@@ -507,7 +507,7 @@ status_t MIndexedList::NodeBefore(int32 index, ArpIndexedNode **answer,
 status_t MIndexedList::NodeAfter(int32 index, ArpIndexedNode **answer,
 						int32 findNode)
 {
-	long hash = HashForIndex(index);
+	int32 hash = HashForIndex(index);
 	if (! HashExists(hash)) return B_ERROR;
 
 	ArpIndexedNode	*nodeAfter = NULL,
@@ -547,7 +547,7 @@ status_t MIndexedList::NodeAfter(int32 index, ArpIndexedNode **answer,
 // the appropriate index.
 status_t MIndexedList::pAddNode(ArpIndexedNode *node)
 {
-	long hash = HashForIndex(node->Index());
+	int32 hash = HashForIndex(node->Index());
 #if 0 	// why did I write this?  Isn't the POINT that if the hash is
 		// too large, then the list grows?
 	if (not HashExists(hash)) {
@@ -604,7 +604,7 @@ status_t MIndexedList::pRemoveNode(ArpIndexedNode *node)
 // the nearest node to add it to, so it gets inserted in the list.
 void MIndexedList::pAddNodeAtChainHead(ArpIndexedNode *node, int32 hash)
 {
-	long	k = hash;
+	int32	k = hash;
 	while (++k < count) {
 		if (chain[k] != NULL) {
 			chain[k]->AddNode(node);
